content/modules/ROOT/solutions/rhads-ai/rhdevspaces/ai-agent/requirements.txt
argcomplete==2.0.0
attrs==22.1.0
distlib==0.3.6
exceptiongroup==1.0.1
filelock==3.8.2
flake8==6.0.0
iniconfig==1.1.1
mccabe==0.7.0
packaging==21.3
platformdirs==2.6.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
pytest==7.2.0
PyYAML==6.0
toml==0.10.2
tomli==2.0.1
virtualenv==20.17.1
xmltodict==0.13.0
yq==3.1.0
langgraph
langchain_openai
langchain_tavily
tavily-python
python-dotenv
Flask
gunicorn

